Melhorado a nomenclatura e estrutura de projeto
