--- src/main/java/Projeto/Studio/Zero/StudioZero/StudioZeroApplication.java
@@ -1,11 +1,0 @@
-package Projeto.Studio.Zero.StudioZero;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StudioZeroApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(StudioZeroApplication.class, args);
-	}
-}
