--- src/main/java/Projeto/Studio/Zero/StudioZero/applications/dtos/request/UserDeleteRequestDTO.java
@@ -1,16 +1,0 @@
-package Projeto.Studio.Zero.StudioZero.applications.dtos.request;
-
-import jakarta.validation.constraints.NotNull;
-
-public class UserDeleteRequestDTO {
-    @NotNull
-    private Long id;
-
-    public @NotNull Long getId() {
-        return id;
-    }
-
-    public void setId(@NotNull Long id) {
-        this.id = id;
-    }
-}
