--- src/main/java/Projeto/Studio/Zero/StudioZero/applications/dtos/request/UserReadRequestDTO.java
@@ -1,19 +1,0 @@
-package Projeto.Studio.Zero.StudioZero.applications.dtos.request;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import jakarta.validation.constraints.NotNull;
-
-@JsonNaming(PropertyNamingStrategies.LowerCamelCaseStrategy.class)
-public class UserReadRequestDTO {
-    @NotNull
-    private Long id;
-
-    public @NotNull Long getId() {
-        return id;
-    }
-
-    public void setId(@NotNull Long id) {
-        this.id = id;
-    }
-}
