--- src/main/java/Projeto/Studio/Zero/StudioZero/config/SwaggerConfig.java
@@ -1,19 +1,0 @@
-package Projeto.Studio.Zero.StudioZero.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Contact;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI().info(new Info().title("Studio Zero")
-                .description("")
-                .contact(new Contact().name(""))
-                .version("1.0.0"));
-    }
-}
