--- src/main/java/Projeto/Studio/Zero/StudioZero/domain/enums/ApiError.java
@@ -1,6 +1,0 @@
-package Projeto.Studio.Zero.StudioZero.domain.enums;
-
-public enum ApiError {
-    NOT_FOUND,
-    BAD_REQUEST
-}
