--- src/main/java/Projeto/Studio/Zero/StudioZero/domain/repositories/UserRepository.java
@@ -1,6 +1,0 @@
-package Projeto.Studio.Zero.StudioZero.domain.repositories;
-
-import Projeto.Studio.Zero.StudioZero.domain.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {}
