+++ src/main/java/com/studiozero/projeto/StudioZeroApplication.java
@@ -1,0 +1,11 @@
+package com.studiozero.projeto;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StudioZeroApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(StudioZeroApplication.class, args);
+	}
+}
