+++ src/main/java/com/studiozero/projeto/domain/enums/ApiError.java
@@ -1,0 +1,6 @@
+package com.studiozero.projeto.domain.enums;
+
+public enum ApiError {
+    NOT_FOUND,
+    BAD_REQUEST
+}
