+++ src/main/java/com/studiozero/projeto/domain/exceptions/ApiException.java
@@ -1,0 +1,14 @@
+package com.studiozero.projeto.domain.exceptions;
+
+import com.studiozero.projeto.domain.enums.ApiError;
+
+public abstract class ApiException extends RuntimeException {
+
+    public ApiException(String message) {
+        super(message);
+    }
+
+    public abstract ApiError apiError();
+    public abstract String userResponseMessage();
+    public abstract int httpStatus();
+}
