+++ src/main/java/com/studiozero/projeto/domain/exceptions/BadRequestException.java
@@ -1,0 +1,26 @@
+package com.studiozero.projeto.domain.exceptions;
+
+import com.studiozero.projeto.domain.enums.ApiError;
+import org.springframework.http.HttpStatus;
+
+public class BadRequestException extends ApiException {
+
+    public BadRequestException(String message) {
+        super(message);
+    }
+
+    @Override
+    public ApiError apiError() {
+        return ApiError.BAD_REQUEST;
+    }
+
+    @Override
+    public String userResponseMessage() {
+        return getMessage();
+    }
+
+    @Override
+    public int httpStatus() {
+        return HttpStatus.BAD_REQUEST.value();
+    }
+}
