+++ src/main/java/com/studiozero/projeto/domain/repositories/UserRepository.java
@@ -1,0 +1,6 @@
+package com.studiozero.projeto.domain.repositories;
+
+import com.studiozero.projeto.domain.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {}
