+++ src/main/java/com/studiozero/projeto/domain/services/UserService.java
@@ -1,0 +1,21 @@
+package com.studiozero.projeto.domain.services;
+
+import com.studiozero.projeto.domain.entities.User;
+import com.studiozero.projeto.applications.web.dtos.request.UserCreateRequestDTO;
+import com.studiozero.projeto.applications.web.dtos.request.UserDeleteRequestDTO;
+import com.studiozero.projeto.applications.web.dtos.request.UserReadRequestDTO;
+import com.studiozero.projeto.applications.web.dtos.request.UserUpdateRequestDTO;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserService {
+    User save(UserCreateRequestDTO request);
+    List<User> searchAll();
+    Optional<User> read(UserReadRequestDTO request);
+    User update(UserUpdateRequestDTO request);
+    Optional<User> delete(UserDeleteRequestDTO request);
+}
+
+
+
