--- src/test/java/Projeto/Studio/Zero/StudioZero/StudioZeroApplicationTests.java
@@ -1,13 +1,0 @@
-package Projeto.Studio.Zero.StudioZero;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class StudioZeroApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
