+++ src/test/java/com/studiozero/projeto/StudioZeroApplicationTests.java
@@ -1,0 +1,13 @@
+package com.studiozero.projeto;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class StudioZeroApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
